requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
sqlalchemy[asyncio]==2.0.25
asyncpg==0.27.0 
redis==5.0.1
fastapi-cache2==0.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic-settings==2.2.1
alembic==1.13.1